+++ src/refactoring/App.tsx
@@ -1,9 +1,11 @@
+import CustomButton from './components/CustomButton.tsx';
+
 import { useState } from 'react';
-import { CartPage } from './components/CartPage.tsx';
-import { AdminPage } from './components/AdminPage.tsx';
+
 import { Coupon, Product } from '../types.ts';
 import { useCoupons, useProducts } from './hooks';
-import CustomButton from './components/CustomButton.tsx';
+import { AdminPage } from './pages/AdminPage.tsx';
+import { CartPage } from './pages/CartPage.tsx';
 
 const initialProducts: Product[] = [
   {
